кидаю бедешку муа

+++ menu/models.py
@@ -4,7 +4,7 @@ from django.db import models
 
 class Dish(models.Model):
     name = models.CharField(max_length=50, verbose_name="Название блюда", null=False)
-    price = models.FloatField(verbose_name="Цена в рублях", null=False)
+    price = models.IntegerField(verbose_name="Цена в рублях", null=False)
     additional_info = models.CharField(max_length=20, verbose_name="Количество (гр, шт, мл)", null=True, blank=True )
     compound = models.TextField(verbose_name="Состав", null=True, blank=True)
     picture = models.ImageField(verbose_name="Фото блюда", null=True, upload_to='dishes/', default="/")
